requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   pynacl
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.5
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   edx-django-utils
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
edx-django-utils==5.4.0
    # via -r requirements/base.txt
exceptiongroup==1.1.1
    # via pytest
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   xblock
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
lxml==4.9.2
    # via
    #   -r requirements/base.txt
    #   xblock
markupsafe==2.1.2
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   xblock
newrelic==8.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==23.1
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
psutil==5.9.5
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.3.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   xblock
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
    #   xblock
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   xblock
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
web-fragments==2.0.0
    # via
    #   -r requirements/base.txt
    #   xblock
webob==1.8.7
    # via
    #   -r requirements/base.txt
    #   xblock
xblock==1.6.2
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
